test: add test cases for isSubDirOrEqual
